Issue #11: Segregated the model structure
- all the model architectures will reside in src/models.py
- all the data processing methods will reside in src/data.py
- all the evaluation methods will reside in src/evaluation.py
- src/sch_imp_nn_exp3.py would use the above helper methods for training
the model and evaluating the same
